Редактирование 4 примера (деления)

+++ Examole003/Program.cs
@@ -1,12 +1,12 @@
 ﻿Console.Write("Введите первое число ");
 
-int numberA = Convert.ToInt32(Console.ReadLine());
+Double numberA = Convert.ToDouble(Console.ReadLine());
 
 Console.Write("Введите второе число ");
 
-int numberB = Convert.ToInt32(Console.ReadLine());
+Double numberB = Convert.ToDouble(Console.ReadLine());
 
-int result = numberA + numberB;
+double result = numberA + numberB;
 
 Console.Write("Сумма чисел ровна ");
